+++ app/src/main/java/com/palmnote/data/db/entity/Budget.kt
@@ -10,7 +10,7 @@ data class Budget(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
     val yearMonth: String, // "2024-01"
-    val totalBudget: Double,
+    val totalBudget: Long, // 预算总额（分）
     val reminderThreshold: Double = 0.8, // 80%
     val reminderAtThreshold: Boolean = true,
     val reminderOverBudget: Boolean = true,
